requirements.txt
Django>=6.0.0
djangorestframework>=3.14.0
django-extensions>=3.2.0
django-debug-toolbar>=4.2.0
phonenumber_field>=7.1.0
python-dotenv>=1.0.0
